% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.bm.exactloglikelihood}
\alias{ds.bm.exactloglikelihood}
\title{Exact calculation of the log-likelihood for a Boltzmann machine}
\usage{
ds.bm.exactloglikelihood(datasources, bm, data = "D")
}
\arguments{
\item{datasources}{A list of Opal object(s) as a handle to the server-side session}

\item{bm}{The name of the Boltzmann machine model on the server-side.}

\item{data}{The name of the variable that holds the data on the server-side.
Defaults to \code{"D"}.}
}
\description{
Calculates the log-likelihood of a Boltzmann machine model.
}
\details{
\emph{This is only feasible for very small models, as the runtime grows exponentially with
the number of hidden nodes.}
}
