% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.dbm.exactloglikelihood}
\alias{ds.dbm.exactloglikelihood}
\title{Exact calculation of the log-likelihood of a DBM}
\usage{
ds.dbm.exactloglikelihood(datasources, dbm = "dbm", data = "D")
}
\arguments{
\item{dbm}{The name of the model to sample from on the server-side. Defaults to \code{"dbm"}.}
}
\description{
Same as \code{\link{ds.bm.exactloglikelihood}}, only with parameter \code{bm} changed to \code{dbm}.
}
