% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.dbm.loglikelihood}
\alias{ds.dbm.loglikelihood}
\title{Likelihood estimation for a DBM model}
\usage{
ds.dbm.loglikelihood(datasources, dbm = "dbm", data = "D",
  parallelized = NULL, ntemperatures = NULL, nparticles = NULL,
  burnin = NULL)
}
\arguments{
\item{datasources}{A list of Opal object(s) as a handle to the server-side session}

\item{dbm}{The name of the DBM model on the server-side. Defaults to \code{"dbm"}.}

\item{data}{The name of the variable that holds the data on the server-side.
Defaults to \code{"D"}.}

\item{ntemperatures}{Number of temperatures for annealing from the starting model
to the target model, defaults to 100}

\item{nparticles}{Number of parallel chains and calculated weights in AIS, defaults to 100}

\item{burnin}{Number of steps to sample for the Gibbs transition between models in AIS}
}
\description{
Performs a separate run of annealed importance sampling (AIS)
for each of the samples to estimate the log-likelihood of a DBM in addition to
estimating the partition function.
}
