% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.dbm.top2LatentDims}
\alias{ds.dbm.top2LatentDims}
\title{Two-dimensional representation of latent features}
\usage{
ds.dbm.top2LatentDims(datasources, dbm = "dbm", data = "D")
}
\arguments{
\item{datasources}{A list of Opal object(s) as a handle to the server-side session}

\item{dbm}{The name of the DBM model on the server-side. Defaults to \code{"dbm"}.}

\item{data}{The name of the variable that holds the data on the server-side.
Defaults to \code{"D"}.}
}
\value{
A matrix with two columns for each of the sites,
  containing the two-dimensional representation for
  each of the samples in the rows. The samples are shuffled at random.
}
\description{
Calculates the mean-field approximation of the activation of the top hidden nodes in
the DBM. If there are more than two nodes in the top layer, the activation is further
reduced in dimensionality using a PCA.
}
