% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.rbm.loglikelihood}
\alias{ds.rbm.loglikelihood}
\title{Likelihood  estimation for an RBM model}
\usage{
ds.rbm.loglikelihood(datasources, rbm = "rbm", data = "D",
  parallelized = NULL, ntemperatures = NULL, nparticles = NULL,
  burnin = NULL)
}
\arguments{
\item{datasources}{A list of Opal object(s) as a handle to the server-side session}

\item{rbm}{The name of the RBM model on the server. Defaults to \code{"rbm"}.}

\item{data}{The name of the variable that holds the data on the server-side.
Defaults to \code{"D"}.}

\item{ntemperatures}{Number of temperatures for annealing from the starting model
to the target model, defaults to 100}

\item{nparticles}{Number of parallel chains and calculated weights in AIS, defaults to 100}

\item{burnin}{Number of steps to sample for the Gibbs transition between models in AIS}
}
\description{
Estimates the log-likelihood for an RBM model using annealed importance sampling (AIS)
for estimating the partition function.
}
