% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.rbm.samples}
\alias{ds.rbm.samples}
\title{Generate samples from a restricted Boltzmann machine}
\usage{
ds.rbm.samples(datasources, rbm = "rbm", ...)
}
\arguments{
\item{rbm}{The name of the model to sample from on the server-side. Defaults to \code{"rbm"}.}
}
\description{
Same as \code{\link{ds.bm.samples}}, only with parameter \code{bm} changed to \code{rbm}.
}
