% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.setJuliaSeed}
\alias{ds.setJuliaSeed}
\title{Set a random seed in Julia}
\usage{
ds.setJuliaSeed(datasources, seed)
}
\arguments{
\item{seed}{the seed to set}
}
\description{
Set a seed for the random generator in Julia.
This makes calls to the non-deterministic algorithms in this package reproducible.
}
