% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ds.splitdata}
\alias{ds.splitdata}
\title{Split samples of a data set}
\usage{
ds.splitdata(datasources, data, ratio, newobj1, newobj2)
}
\arguments{
\item{datasources}{A list of Opal object(s) as a handle to the server-side session}

\item{data}{The name of the variable that holds the data on the server-side. Defaults to \code{"D"}.}

\item{ratio}{The ratio of samples}

\item{newobj1}{The name for the variable where \eqn{x_1} is stored}

\item{newobj2}{The name for the variable where \eqn{x_2} is stored}
}
\description{
Splits a data set randomly into two data sets \eqn{x_1} and \eqn{x_2}, such that
the fraction of samples in \eqn{x_2} is equal to (or as close as possible to) the
given \code{ratio}.
}
\examples{
    ds.splitdata(o, "D", 0.1, "D.Train", "D.Test")
}
